serena/pyproject.toml
[project]
name = "serena"
version = "0.1.0"
description = "Task Management AI Agent using Google ADK"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-adk>=1.0.0",
    "litellm",
    "python-dotenv",
    "uvicorn",
]
